reduced duplication between casperjs and phantomjs snap files
