+++ lib/wraith/javascript/_phantom__common.js
@@ -1,7 +1,8 @@
 module.exports = function (config) {
 
   // modules
-  var page = require('webpage').create();
+  var page   = require('webpage').create(),
+      helper = require('./_helper.js')(systemArgs[2]);
 
   // config
   var systemArgs        = config.systemArgs,
@@ -9,7 +10,7 @@ module.exports = function (config) {
 
   // command line arguments
   var url = systemArgs[1],
-      dimensions = require('./_getDimensions.js')(systemArgs[2]),
+      dimensions = helper.dimensions,
       image_name = systemArgs[3],
       selector   = systemArgs[4],
       globalBeforeCaptureJS = systemArgs[5],
@@ -24,9 +25,9 @@ module.exports = function (config) {
   var last_request_timeout;
   var final_timeout;
 
-  if (takingMultipleScreenshots()) {
+  if (helper.takingMultipleScreenshots()) {
     currentDimensions = dimensions[0];
-    image_name = replaceImageNameWithDimensions(image_name, currentDimensions);
+    image_name = helper.replaceImageNameWithDimensions(image_name, currentDimensions);
   }
   else {
     currentDimensions = dimensions;
@@ -37,26 +38,6 @@ module.exports = function (config) {
 
   // If you want to use additional phantomjs commands, place them here
   page.settings.userAgent = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/28.0.1500.95 Safari/537.17';
-
-  // You can place custom headers here, example below.
-  // page.customHeaders = {
-
-  //      'X-Candy-OVERRIDE': 'https://api.live.bbc.co.uk/'
-
-  //  };
-
-  // If you want to set a cookie, just add your details below in the following way.
-
-  // phantom.addCookie({
-  //     'name': 'ckns_policy',
-  //     'value': '111',
-  //     'domain': '.bbc.co.uk'
-  // });
-  // phantom.addCookie({
-  //     'name': 'locserv',
-  //     'value': '1#l1#i=6691484:n=Oxford+Circus:h=e@w1#i=8:p=London@d1#1=l:2=e:3=e:4=2@n1#r=40',
-  //     'domain': '.bbc.co.uk'
-  // });
 
   page.onResourceRequested = function(req) {
     current_requests += 1;
@@ -75,25 +56,6 @@ module.exports = function (config) {
       phantom.exit();
     }
   });
-
-  function takingMultipleScreenshots() {
-    return dimensions.length && dimensions.length > 1;
-  }
-
-  function replaceImageNameWithDimensions(image_name, currentDimensions) {
-    // shots/clickable_guide__after_click/MULTI_casperjs_english.png
-    // ->
-    // shots/clickable_guide__after_click/1024x359_casperjs_english.png
-    var dirs = image_name.split('/'),
-        filename = dirs[dirs.length - 1],
-        filenameParts = filename.split('_'),
-        newFilename;
-
-    filenameParts[0] = currentDimensions.viewportWidth + 'x' + currentDimensions.viewportHeight;
-    dirs.pop(); // remove MULTI_casperjs_english.png
-    newFilename = dirs.join('/') + '/' + filenameParts.join('_');
-    return newFilename;
-  }
 
   function runSetupJavaScriptThenCaptureImage() {
     if (globalBeforeCaptureJS) {
@@ -117,9 +79,9 @@ module.exports = function (config) {
     page.render(image_name);
 
     dimensionsProcessed++;
-    if (takingMultipleScreenshots() && dimensionsProcessed < dimensions.length) {
+    if (helper.takingMultipleScreenshots() && dimensionsProcessed < dimensions.length) {
       currentDimensions = dimensions[dimensionsProcessed];
-      image_name = replaceImageNameWithDimensions(image_name, currentDimensions);
+      image_name = helper.replaceImageNameWithDimensions(image_name, currentDimensions);
       setTimeout(captureImage, 1000);
     }
     else {
