+++ lib/wraith/javascript/casper.js
@@ -1,10 +1,11 @@
 // modules
 var system = require('system'),
-    casper = require('casper').create();
+    casper = require('casper').create(),
+    helper = requireRelative('_helper.js')(casper.cli.get(1));
 
 // command line arguments
 var url = casper.cli.get(0),
-    dimensions = requireRelative('_getDimensions.js')(casper.cli.get(1)),
+    dimensions = helper.dimensions,
     image_name = casper.cli.get(2),
     selector = casper.cli.get(3),
     globalBeforeCaptureJS = casper.cli.get(4),
@@ -13,9 +14,6 @@ var url = casper.cli.get(0),
     currentDimensions;
 
 // functions
-function takingMultipleScreenshots() {
-  return dimensions.length && dimensions.length > 1;
-}
 function requireRelative(file) {
   // PhantomJS will automatically `require` relatively, but CasperJS needs some extra help. Hence this function.
   // 'templates/javascript/casper.js' -> 'templates/javascript'
@@ -36,33 +34,19 @@ function snap() {
   }
 
   dimensionsProcessed++;
-  if (takingMultipleScreenshots() && dimensionsProcessed < dimensions.length) {
+  if (helper.takingMultipleScreenshots() && dimensionsProcessed < dimensions.length) {
     currentDimensions = dimensions[dimensionsProcessed];
-    image_name = replaceImageNameWithDimensions(image_name, currentDimensions);
+    image_name = helper.replaceImageNameWithDimensions(image_name, currentDimensions);
     casper.viewport(currentDimensions.viewportWidth, currentDimensions.viewportHeight);
     casper.wait(300, function then () {
       snap.bind(this)();
     });
   }
 }
-function replaceImageNameWithDimensions(image_name, currentDimensions) {
-  // shots/clickable_guide__after_click/MULTI_casperjs_english.png
-  // ->
-  // shots/clickable_guide__after_click/1024x359_casperjs_english.png
-  var dirs = image_name.split('/'),
-      filename = dirs[dirs.length - 1],
-      filenameParts = filename.split('_'),
-      newFilename;
 
-  filenameParts[0] = currentDimensions.viewportWidth + 'x' + currentDimensions.viewportHeight;
-  dirs.pop(); // remove MULTI_casperjs_english.png
-  newFilename = dirs.join('/') + '/' + filenameParts.join('_');
-  return newFilename;
-}
-
-if (takingMultipleScreenshots()) {
+if (helper.takingMultipleScreenshots()) {
   currentDimensions = dimensions[0];
-  image_name = replaceImageNameWithDimensions(image_name, currentDimensions);
+  image_name = helper.replaceImageNameWithDimensions(image_name, currentDimensions);
 }
 else {
   currentDimensions = dimensions;
